feat: add type value as optional configuration
javascript-time-ago offers a style attribute to group time differently.
By adding this option, we enable that grouping in Ilenia by still
setting the same default behaviour as before so no current usage
should be broken.
Reason for changing name is that style is a react prop and I didn't want
to confuse them together as they serve different purpose.